+++ CHANGELOG.md
@@ -14,5 +14,6 @@
 
 ### Maintenance
 
+- 2026-09-19 | 🔧 chore(server): prepare hosted configuration
 - 2026-09-19 | 🔧 chore(server): validate Vapor bootstrap
 - 2026-09-19 | 🔧 chore: initialize iOS and Vapor workspace